Python Django/data-analysis/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.1
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.14.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.6
kombu==5.3.5
numpy==1.26.3
packaging==23.2
pandas==2.1.4
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
stripe==8.0.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.25.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
